requirements.txt
# Core web framework
flask==3.0.0
flask-cors==4.0.0

# Zoom SDK
zoomus==1.1.9

# Audio processing and transcription
openai-whisper==20231117
openai==1.3.0
anthropic==0.7.0

# Alternative transcription service
deepgram-sdk==3.0.0

# Audio/Video processing
pydub==0.25.1
sounddevice==0.4.6
soundfile==0.12.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Environment variables
python-dotenv==1.0.0

# Async support
aiohttp==3.9.1

# Zoom Meeting SDK (for browser-less bot)
selenium==4.15.2
webdriver-manager==4.0.1

# Meeting bot alternative - using Recall.ai API (recommended)
requests==2.31.0

# Job queue for async meeting processing
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
